Split product data setup out, use product query
# Conflicts:
#	src/RemoteInboxNotifications/RemoteInboxNotificationsEngine.php
