+++ src/RemoteInboxNotifications/RemoteInboxNotificationsEngine.php
@@ -26,8 +26,7 @@ class RemoteInboxNotificationsEngine {
 	public static function init() {
 		add_action( 'activated_plugin', array( __CLASS__, 'run' ) );
 		add_action( 'deactivated_plugin', array( __CLASS__, 'run_on_deactivated_plugin' ), 10, 1 );
-		add_action( 'product_page_product_importer', array( __CLASS__, 'run_on_product_importer' ) );
-		add_action( 'transition_post_status', array( __CLASS__, 'run_on_transition_post_status' ), 10, 3 );
+		DataSetupForProducts::init();
 
 		// Pre-fetch data so it has the correct initial values.
 		self::get_data();
@@ -57,14 +56,13 @@ class RemoteInboxNotificationsEngine {
 	 *
 	 * @return object The data option.
 	 */
-	private static function get_data() {
+	public static function get_data() {
 		$data = get_option( self::DATA_OPTION_NAME );
 
 		if ( false === $data ) {
-			$data                         = new \stdClass();
-			$products_provider            = new ProductsProvider();
-			$data->there_were_no_products = ! $products_provider->are_there_products();
-			$data->there_are_now_products = ! $data->there_were_no_products;
+			$data = new \stdClass();
+
+			DataSetupForProducts::init_data( $data );
 
 			add_option( self::DATA_OPTION_NAME, $data );
 		}
@@ -82,53 +80,6 @@ class RemoteInboxNotificationsEngine {
 	 */
 	public static function run_on_deactivated_plugin( $plugin ) {
 		PluginsProvider::set_deactivated_plugin( $plugin );
-		self::run();
-	}
-
-	/**
-	 * Runs on product importer steps.
-	 */
-	public static function run_on_product_importer() {
-		// We're only interested in when the importer completes.
-		// phpcs:disable WordPress.Security.NonceVerification.Recommended
-		if ( ! isset( $_REQUEST['step'] ) ) {
-			return;
-		}
-		if ( 'done' !== $_REQUEST['step'] ) {
-			return;
-		}
-		// phpcs:enable
-
-		$data                         = self::get_data();
-		$products_provider            = new ProductsProvider();
-		$data->there_are_now_products = $products_provider->are_there_products();
-		update_option( self::DATA_OPTION_NAME, $data );
-
-		self::run();
-	}
-
-	/**
-	 * Runs when a post status transitions, but we're only interested if it is
-	 * a product being published.
-	 *
-	 * @param string $new_status The new status.
-	 * @param string $old_status The old status.
-	 * @param Post   $post       The post.
-	 */
-	public static function run_on_transition_post_status( $new_status, $old_status, $post ) {
-		if (
-			'product' !== $post->post_type ||
-			'publish' !== $new_status ||
-			'publish' === $old_status
-		) {
-			return;
-		}
-
-		$data                         = self::get_data();
-		$products_provider            = new ProductsProvider();
-		$data->there_are_now_products = $products_provider->are_there_products();
-		update_option( self::DATA_OPTION_NAME, $data );
-
 		self::run();
 	}
 }
